# api/swagger.yml
openapi: 3.1.0
info:
  title: Receipt Wrangler API.
  termsOfService:
    there are no TOS at this moment, use at your own risk we take no
    responsibility
  license:
    name: GPLV3
    url: https://www.gnu.org/licenses/gpl-3.0.en.html
  version: 5.0.0
servers:
  - url: /api
security:
  - bearerAuth: [ ]
  - apiKeyAuth: [ ]
paths:
...
  /user:
    get:
      tags:
        - User
      summary: Get users
      description:
        This will get all the users in the system and return a view without
        sensative information
      operationId: getUsers
      responses:
        200:
          description: The users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserView"
        500:
          $ref: "#/components/responses/Internal"
      security:
        - bearerAuth: [ ]
        - apiKeyAuth: [ ]
    post:
      tags:
        - User
      summary: Create user
      description: This will to create a user, [SYSTEM ADMIN]
      requestBody:
        description: User to create
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      operationId: createUser
      responses:
        200:
          $ref: "#/components/responses/Ok"
        500:
          $ref: "#/components/responses/Internal"
      security:
        - bearerAuth: [ ]
        - apiKeyAuth: [ ]
...
  /user/{username}:
    get:
      tags:
        - User
      summary: Get username count
      description:
        This will return the number of users in the system with the same
        username
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: Username to get the count of
      operationId: getUsernameCount
      responses:
        200:
          description: The number of users with the same username
          content:
            application/json:
              schema:
                type: integer
        500:
          $ref: "#/components/responses/Internal"
  /user/amountOwedForUser:
    get:
      tags:
        - User
      summary: Get amount owed for user
      description:
        This will return the amount owed for the logged in user, in the
        specified group, [SYSTEM USER]
      parameters:
        - in: query
          name: groupId
          schema:
            type: integer
          required: false
          description: The Id of the group to get amount owed for
        - in: query
          name: receiptIds
          schema:
            type: array
            items:
              type: integer
          required: false
          description: The Id of the receipts to get amount owed for
      operationId: getAmountOwedForUser
      responses:
        200:
          description: The amount owed
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
        500:
          $ref: "#/components/responses/Internal"
      security:
        - bearerAuth: [ ]
        - apiKeyAuth: [ ]
...
  /user/updateUserProfile:
    put:
      tags:
        - User
      summary: Update user profile
      description: This will update the logged in user's user profile
      requestBody:
        description: User profile to update
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileCommand"
      operationId: updateUserProfile
      responses:
        200:
          $ref: "#/components/responses/Ok"
        500:
          $ref: "#/components/responses/Internal"
      security:
        - bearerAuth: [ ]
        - apiKeyAuth: [ ]
  /user/deleteAccount:
    post:
      tags:
        - User
      summary: Delete own account
      description: This will delete the currently logged in user's account after verifying their password
      requestBody:
        description: Password confirmation for account deletion
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteAccountCommand"
      operationId: deleteAccount
      responses:
        200:
          $ref: "#/components/responses/Ok"
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/Unauthorized"
        500:
          $ref: "#/components/responses/Internal"
      security:
        - bearerAuth: [ ]
        - apiKeyAuth: [ ]
  /user/appData:
    get:
      tags:
        - User
      summary: Get app data
      description:
        This will return the user's app data for the currently logged in
        user [SYSTEM USER]
      operationId: getAppData
      responses:
        200:
          description: The user's app data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppData"
        403:
          $ref: "#/components/responses/Forbidden"
        500:
          $ref: "#/components/responses/Internal"
      security:
        - bearerAuth: [ ]
        - apiKeyAuth: [ ]
  /user/bulk:
    delete:
      tags:
        - User
      summary: Bulk delete users
      description: This will delete multiple users by their IDs [SYSTEM ADMIN]
      requestBody:
        description: User IDs to delete
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkUserDeleteCommand"
      operationId: bulkDeleteUsers
      responses:
        200:
          $ref: "#/components/responses/Ok"
        400:
          $ref: "#/components/responses/BadRequest"
        403:
          $ref: "#/components/responses/Forbidden"
        500:
          $ref: "#/components/responses/Internal"
      security:
        - bearerAuth: [ ]
        - apiKeyAuth: [ ]
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: API Key in format v1.xxx.xxx.xxx
  responses:
    Ok:
      description: Request was successfully processed
    BadRequest:
      description: The request was malformed
    Forbidden:
      description: The request was not allowed
    Unauthorized:
      description: The request was not authorized
    Internal:
      description: There was an error processing the request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InternalErrorResponse'
  schemas:
    ReceiptStatus:
      description: Status of a receipt
      type: string
      enum:
        - "OPEN"
        - "NEEDS_ATTENTION"
        - "RESOLVED"
        - "DRAFT"
        - ""
...
    GroupStatus:
      type: string
      enum:
        - "ACTIVE"
        - "ARCHIVED"
...
    UserRole:
      type: string
      enum:
        - ADMIN
        - USER
...
    CurrencySeparator:
      type: string
      enum:
        - ","
        - "."
    CurrencySymbolPosition:
      type: string
      enum:
        - "START"
        - "END"
...
    Category:
      type: object
      properties:
        createdAt:
          type: string
        createdBy:
          type: integer
          format: uint64
        id:
          type: integer
          format: uint64
        name:
          type: string
          description: Name of the category
        description:
          type: string
          description: Description of the category
        updatedAt:
          type: string
          x-go-name: UpdatedAt
      description: Category to relate receipts to
...
    Group:
      required:
        - name
        - status
        - id
        - groupMembers
        - isAllGroup
        - groupReceiptSettings
      type: object
      properties:
        createdAt:
          type: string
        createdBy:
          type: integer
          format: uint64
        groupSettings:
          $ref: "#/components/schemas/GroupSettings"
        groupReceiptSettings:
          $ref: "#/components/schemas/GroupReceiptSettings"
        groupMembers:
          type: array
          description: Members of the group
          items:
            $ref: "#/components/schemas/GroupMember"
        id:
          type: integer
          format: uint64
        isDefault:
          type: boolean
          description: Is default group (not used yet)
        name:
          type: string
          description: Name of the group
        isAllGroup:
          type: boolean
          description: Is all group for user
        status:
          type: string
          $ref: "#/components/schemas/GroupStatus"
        updatedAt:
          type: string
      description: Group in the system
...
    Tag:
      required:
        - name
      type: object
      properties:
        createdAt:
          type: string
          x-go-name: CreatedAt
        createdBy:
          type: integer
          format: uint64
        id:
          type: integer
          format: uint64
        name:
          type: string
          description: Tag name
        description:
          type: string
          description: Tag description
        updatedAt:
          type: string
      description: Tag to relate receipts to
...
    User:
      required:
        - username
        - displayName
        - isDummyUser
        - userRole
        - id
      type: object
      properties:
        password:
          type: string
          description: User's password
        username:
          type: string
          description: User's username used to login
        createdAt:
          type: string
        createdBy:
          type: integer
          format: uint64
        defaultAvatarColor:
          type: string
          description: Default avatar color
          x-go-name: DefaultAvatarColor
        displayName:
          type: string
          description: Display name
          x-go-name: DisplayName
        id:
          type: integer
        isDummyUser:
          type: boolean
          description: Is dummy user
          x-go-name: IsDummyUser
        updatedAt:
          type: string
          x-go-name: UpdatedAt
        userRole:
          type: string
          $ref: "#/components/schemas/UserRole"
          description: User's role
        lastLoginDate:
          type: string
      description: User in the system
      x-go-package: receipt-wrangler/api/internal/models
    UserView:
      required:
        - username
        - displayName
        - isDummyUser
        - userRole
        - id
      type: object
      properties:
        username:
          type: string
          description: User's username used to login
        createdAt:
          type: string

        createdBy:
          type: integer
          format: uint64
        defaultAvatarColor:
          type: string
          description: Default avatar color
        displayName:
          type: string
          description: Display name
        id:
          type: integer
        isDummyUser:
          type: boolean
          description: Is dummy user
        updatedAt:
          type: string
        userRole:
          type: string
          $ref: "#/components/schemas/UserRole"
          description: User's role
      description: User in the system
      x-go-package: receipt-wrangler/api/internal/models
    UpdateProfileCommand:
      required:
        - displayName
        - defaultAvatarColor
      type: object
      properties:
        displayName:
          type: string
          description: User's displayName
        defaultAvatarColor:
          type: string
          description: Color of default avatar
      description: Command to update user's profile
...
    DeleteAccountCommand:
      required:
        - password
      type: object
      properties:
        password:
          type: string
          description: User's current password for confirmation
      description: Command to delete the user's own account
...
    BulkUserDeleteCommand:
      required:
        - userIds
      type: object
      properties:
        userIds:
          type: array
          items:
            type: string
          description: User IDs to delete
...
    FeatureConfig:
      required:
        - enableLocalSignUp
        - aiPoweredReceipts
      type: object
      properties:
        aiPoweredReceipts:
          type: boolean
          description: Whether AI powered receipts are enabled
        enableLocalSignUp:
          type: boolean
          description: Whether local sign up is enabled
...
    UserPreferences:
      allOf:
        - $ref: "#/components/schemas/BaseModel"
        - type: object
          required:
            - id
            - userId
          properties:
            id:
              type: integer
              description: User preferences id
            userId:
              type: integer
              description: User foreign key
            quickScanDefaultGroupId:
              type: integer
              description: Group foreign key
              default: 0
            quickScanDefaultPaidById:
              type: integer
              description: User foreign key
              default: 0
            quickScanDefaultStatus:
              type: string
              $ref: "#/components/schemas/ReceiptStatus"
              description: Default quick scan status
              default: "OPEN"
            showLargeImagePreviews:
              type: boolean
              description: Whether to show large image previews
              default:
            userShortcuts:
              type: array
              items:
                $ref: "#/components/schemas/UserShortcut"
...
    GroupReceiptSettings:
      allOf:
        - $ref: "#/components/schemas/BaseModel"
        - type: object
          required:
            - groupId
          properties:
            groupId:
              type: integer
              description: Group foreign key
            hideImages:
              type: boolean
              description: Hide receipt images
            hideReceiptCategories:
              type: boolean
              description: Hide receipt categories
            hideReceiptTags:
              type: boolean
              description: Hide receipt tags
            hideItemCategories:
              type: boolean
              description: Hide receipt item categories
            hideItemTags:
              type: boolean
              description: Hide receipt item tags
            hideComments:
              type: boolean
              description: Hide receipt comments
            hideShareCategories:
              type: boolean
              description: Hide share categories
            hideShareTags:
              type: boolean
              description: Hide share tags
...
    GroupSettings:
      type: object
      required:
        - id
        - groupId
      properties:
        id:
          type: integer
          description: Group settings id
        groupId:
          type: integer
          description: Group foreign key
        emailIntegrationEnabled:
          type: boolean
          description: Whether email integration is enabled
        systemEmailId:
          type: integer
          description: System email foreign key
        systemEmail:
          $ref: "#/components/schemas/SystemEmail"
        emailToRead:
          type: string
          description: Email to read
        subjectLineRegexes:
          type: array
          description: Subject line regexes
          items:
            $ref: "#/components/schemas/SubjectLineRegex"
        emailWhiteList:
          type: array
          description: Email white list
          items:
            $ref: "#/components/schemas/GroupSettingsWhiteListEmail"
        emailDefaultReceiptStatus:
          type: string
          $ref: "#/components/schemas/ReceiptStatus"
          description: Default receipt status
        emailDefaultReceiptPaidById:
          type: integer
          description: User foreign key
        prompt:
          $ref: "#/components/schemas/Prompt"
        promptId:
          type: integer
          description: Prompt foreign key
        fallbackPrompt:
          $ref: "#/components/schemas/Prompt"
        fallbackPromptId:
          type: integer
          description: Fallback prompt foreign key
        createdAt:
          type: string
        createdBy:
          type: integer
          format: uint64
        updatedAt:
          type: string
...
    AppData:
      type: object
      required:
        - about
        - claims
        - groups
        - users
        - userPreferences
        - featureConfig
        - categories
        - tags
        - currencyDisplay
        - icons
      properties:
        about:
          $ref: "#/components/schemas/About"
        claims:
          $ref: "#/components/schemas/Claims"
        groups:
          type: array
          description: Groups in the system
          items:
            $ref: "#/components/schemas/Group"
        users:
          type: array
          description: Users in the system
          items:
            $ref: "#/components/schemas/UserView"
        userPreferences:
          $ref: "#/components/schemas/UserPreferences"
        featureConfig:
          $ref: "#/components/schemas/FeatureConfig"
        categories:
          type: array
          description: Categories in the system
          items:
            $ref: "#/components/schemas/Category"
        tags:
          type: array
          description: Tags in the system
          items:
            $ref: "#/components/schemas/Tag"
        jwt:
          type: string
          description: JWT token
        refreshToken:
          type: string
          description: Refresh token
        currencyDisplay:
          type: string
          description: Currency display
        currencyThousandthsSeparator:
          $ref: "#/components/schemas/CurrencySeparator"
        currencyDecimalSeparator:
          $ref: "#/components/schemas/CurrencySeparator"
        currencySymbolPosition:
          $ref: "#/components/schemas/CurrencySymbolPosition"
        currencyHideDecimalPlaces:
          type: boolean
          description: Whether to hide decimal places
        icons:
          type: array
          description: Icons in the system
          items:
            $ref: "#/components/schemas/Icon"
...
    Claims:
      type: object
      required:
        - userId
        - userRole
        - displayName
        - defaultAvatarColor
        - username
        - iss
        - exp
      properties:
        userId:
          type: integer
          description: User foreign key
          default: 0
        userRole:
          type: string
          $ref: "#/components/schemas/UserRole"
          description: User's role
          default: USER
        displayName:
          type: string
          description: Display name
          default: ""
        defaultAvatarColor:
          type: string
          description: Default avatar color
          default: ""
        username:
          type: string
          description: User's username used to login
          default: ""
        iss:
          type: string
          description: Issuer
          default: ""
        sub:
          type: string
          description: Subject
          default: ""
        aud:
          type: array
          items:
            type: string
          description: Audience
          default: [ ]
        exp:
          type: integer
          description: Expiration time
          default: 0
        nbf:
          type: integer
          description: Not before
          default: 0
        iat:
          type: integer
          description: Issued at
          default: 0
        jti:
          type: string
          description: JWT ID
          default: ""
...
    Icon:
      type: object
      required:
        - value
        - displayValue
      properties:
        value:
          type: string
          description: Icon value
        displayValue:
          type: string
          description: Icon display value
    About:
      type: object
      required:
        - buildDate
        - version
      properties:
        buildDate:
          type: string
          description: Build date
        version:
          type: string
          description: Version
...
    InternalErrorResponse:
      type: object
      required:
        - errorMsg
      properties:
        errorMsg:
          type: string
          description: Error message
...